#7058 Use Quantity instead of BigDecimal
